+++ packages/auth/.eslintrc.js
@@ -1,0 +1,8 @@
+/** @type {import("eslint").Linter.Config} */
+module.exports = {
+  root: true,
+  extends: [require.resolve('@turbocell/eslint-config/node')],
+  parserOptions: {
+    project: `${__dirname}/tsconfig.json`,
+  },
+};
