+++ packages/auth/src/next-auth.ts
