+++ packages/auth/src/react.ts
