+++ packages/auth/src/server.ts
@@ -1,0 +1,74 @@
+import DiscordProvider from "next-auth/providers/discord";
+import GoogleProvider from "next-auth/providers/google";
+import FacebookProvider from "next-auth/providers/facebook";
+
+import { PrismaAdapter } from "@auth/prisma-adapter";
+import type { Adapter } from "@auth/core/adapters";
+import { type DefaultSession } from "next-auth";
+import { db } from "@turbocell/db";
+import type { sessionSchema, userSchema, z } from "@turbocell/db";
+import NextAuth from "./next-auth";
+
+export type { Session, DefaultSession as DefaultAuthSession } from "next-auth";
+
+/**
+ * Module augmentation for `next-auth` types. Allows us to add custom properties to the `session`
+ * object and keep type safety.
+ *
+ * @see https://next-auth.js.org/getting-started/typescript#module-augmentation
+ */
+declare module "next-auth" {
+  type User = z.infer<typeof userSchema>
+  interface Session extends z.infer<typeof sessionSchema> {
+    user: DefaultSession["user"] & {
+      id: string;
+    };
+  }
+  interface Profile {
+    id: string;
+  }
+}
+
+// if (!process.env.GITHUB_ID) {
+//   throw new Error('No GITHUB_ID has been provided.');
+// }
+
+const useSecureCookies = process.env.VERCEL_ENV === "production";
+const cookiePrefix = useSecureCookies ? "__Secure-" : "";
+const cookieDomain = useSecureCookies ? "turbocell.vercel.app" : "localhost";
+
+export const {
+  handlers: { GET, POST },
+  auth,
+} = NextAuth({
+  pages: {
+    signIn: "/sign-in",
+  },
+  cookies: {
+    sessionToken: {
+      name: `${cookiePrefix}next-auth.session-token`,
+      options: {
+        httpOnly: true,
+        sameSite: "lax",
+        path: "/",
+        domain: cookieDomain,
+        secure: useSecureCookies,
+      },
+    },
+  },
+  adapter: PrismaAdapter(db) as Adapter,
+  providers: [
+    DiscordProvider({
+      clientId: process.env.DISCORD_CLIENT_ID!,
+      clientSecret: process.env.DISCORD_CLIENT_SECRET!,
+    }),
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID!,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+    }),
+    FacebookProvider({
+      clientId: process.env.FACEBOOK_CLIENT_ID!,
+      clientSecret: process.env.FACEBOOK_CLIENT_SECRET!,
+    }),
+  ],
+});
