+++ packages/auth/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
   },
-  "include": ["."],
+  "include": ["src/**/*.ts"],
   "exclude": ["node_modules"]
 }
